app2/render-map-b/requirements.txt
osmapi
Pillow
requests

